Add optional types & @init
Optional Types: Types prefixed with a ? that will store a value along
with a "present" state. Currently has no safety checks/ensuring you are
only accessing a value if it is present, and currently there are no
plans to implement that feature. I don't think panics are a particularly
good idea to have coming from intrinsics in this language.

@init: Allows for initializing more complex data structures. For
example, structs can be initialized with '@init(<type>,
<item_values>...), and optionals can be initialized with @init(<type>,
<value_or_none>). This feature is essentially functionally equivalent to
struct initialization in many languages, however I could not think of a
syntax similar to how other languages doing it that wouldn't require a
more complex parsing process. Additionally, as I already have done, this
can be used/extended to work for many data types.