Fix generics implementation from other files
Before, if something generic was implemented with a type it couldn't see
from it's file, then that would be an error. Now we save the location of
where the basic types point to for generics so that they will always be
able to be resolved no matter what file has the generic function or
type.