Implement Macros (part of them at least)
With the macro system in barely, I am looking for something less unsafe
compared to C, however without sacraficing a lot of the power that
macros provide. Instead of being simply expanded at compile time, this
implementation of macros works in a way where the macro definitions can
be completely parsed into AST while parsing, and inserted into their
correct place in the code while processing. This is a much cleaner
implementation (in my opinion) as it allow for a lot more useful checks
to be possible, as well as allowing for macros to potentially work with
modules. These may be removed / reworked completely if I discover big
issues in implementation.