Remove use
I don't think 'use' serves much of a purpose anymore, now that there
aren't any file prefixes for functions or anything of the sort. The only
advantage of 'use' that was left was for a little bit of code clarity,
however for the sake of simplifying the language, I don't think this
code clarity is incredibly important.