Default to parsing macro arg as the current syntax
An assumption that greatly reduces the amount of times that it is needed
to specify the syntax type for a macro arg. I am still on fence about
whether I like this feature or not, but in order to keep code semi
readable, I think this is a good change.