Add support for macros in type definitions
While I don't think that having macros that can expand to items is a
good idea (because it results in the macros needed to be expanded prior
to evaluation). However, I do think that allowing types defined by
macros to exist globally is reasonable, it is trivial to process the
macros inside types when referencing them. This is a feature that may be
removed in the future because of it being a bit complex, however for now
I think it is reasonable.